+++ src/util/mod.rs
@@ -3,6 +3,8 @@
 pub mod file_io;
 /// Homopolymer-compression related functions
 pub mod homopolymers;
+/// Quick utility functions for htslib files
+pub mod htslib_quickparse;
 /// Utility functions related to mappings
 pub mod mapping;
 /// Functions for sequence manipulation